commit 28-09

+++ app/src/main/java/com/hrgirdattendanceonline/AttendanceActivity.java
@@ -1,7 +1,9 @@
 package com.hrgirdattendanceonline;
 
+import android.app.AlertDialog;
 import android.app.ProgressDialog;
 import android.content.Context;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.IntentFilter;
 import android.content.SharedPreferences;
@@ -84,6 +86,7 @@ public class AttendanceActivity extends AppCompatActivity implements MFS100Event
     String Url, url_http;
     String android_id, logo;
     String empattDid, flag;
+    String TabID;
 
     int result_match = 0;
 
@@ -325,7 +328,8 @@ public class AttendanceActivity extends AppCompatActivity implements MFS100Event
     public void deviceData()
     {
         android_id = Settings.Secure.getString(getApplicationContext().getContentResolver(), Settings.Secure.ANDROID_ID);
-        txt_att_deviceid.setText(android_id);
+        //txt_att_deviceid.setText(android_id);
+        //GetDeviceName(android_id);
 
         String manufacturer = Build.MANUFACTURER;
         String model = Build.MODEL;
@@ -806,10 +810,11 @@ public class AttendanceActivity extends AppCompatActivity implements MFS100Event
                 try
                 {
                     String leave_url = ""+url_http+""+Url+"/owner/hrmapi/getallempdatadevicewise/?";
-                    String query3 = String.format("deviceid=%s&flag=%s&empdevicearr=%s",
+                    String query3 = String.format("deviceid=%s&flag=%s&empdevicearr=%s&offline_flag=%s",
                             URLEncoder.encode(android_id, "UTF-8"),
                             URLEncoder.encode(flag, "UTF-8"),
-                            URLEncoder.encode(empattDid, "UTF-8"));
+                            URLEncoder.encode(empattDid, "UTF-8"),
+                            URLEncoder.encode("0", "UTF-8"));
 
                     query3 = query3.replace("%2C+",",");
                     URL url = new URL(leave_url + query3);
@@ -891,6 +896,8 @@ public class AttendanceActivity extends AppCompatActivity implements MFS100Event
                 {
                     myJson2 = result;
                     Log.i("myJson", myJson2);
+
+                    GetDeviceName(android_id);
 
                     progressDialog.dismiss();
 
@@ -1038,7 +1045,9 @@ public class AttendanceActivity extends AppCompatActivity implements MFS100Event
                                         }
 
                                         //Log.i("Insert: ", "Inserting ..");
-                                        db.UpdateContactAttType(new UserDetails_Model(t1,t2,t3,t4,get_attType), get_uId);
+                                        if (db.checkEmpId(get_uId)) {
+                                            db.UpdateContactAttType(new UserDetails_Model(t1, t2, t3, t4, get_attType), get_uId);
+                                        }
                                     }
                                     else if (get_status.equals("3"))
                                     {
@@ -1047,8 +1056,9 @@ public class AttendanceActivity extends AppCompatActivity implements MFS100Event
 
                                         String get_mobile = object.getString("mobile");
                                         //Log.i("get_mobile",get_mobile);
-
-                                        db.deleteContact(get_mobile);
+                                        if (db.checkEmpId(get_uId)) {
+                                            db.deleteContact(get_mobile);
+                                        }
                                     }
                                 }
 
@@ -1092,6 +1102,111 @@ public class AttendanceActivity extends AppCompatActivity implements MFS100Event
 
         GetUserData getUrlData = new GetUserData();
         getUrlData.execute();
+    }
+
+    public void MatchThumbByDB(final FingerData fingerData)
+    {
+        runOnUiThread(new Runnable()
+        {
+            @Override
+            public void run()
+            {
+                List<UserDetails_Model> contacts = db.getAllContacts();
+                Log.i("MFS_Log contacts", "" + contacts);
+
+                result_match = 0;
+
+                /*for (UserDetails_Model cn : contacts)
+                {
+                    Log.i("thumb_reg", cn.getThumb1()+", \n"+cn.getThumb2()+
+                            ", \n"+cn.getThumb3()+", \n"+cn.getThumb4());
+                }*/
+
+                for (UserDetails_Model cn : contacts)
+                {
+                    if (result_match < 1400)
+                    {
+                        String t1 = cn.getThumb1();
+                        String t2 = cn.getThumb2();
+                        String t3 = cn.getThumb3();
+                        String t4 = cn.getThumb4();
+
+                        String thumbs[] = {t1, t2, t3, t4};
+                        Log.i("thumbs",""+  Arrays.toString(thumbs));
+
+                        for (int i = 0; i < thumbs.length; i++)
+                        {
+                            if (result_match < 1400)
+                            {
+                                RegisteredBase64 = thumbs[i];
+                                if (RegisteredBase64 != null)
+                                {
+                                    Log.i("MFS_ RegisteredBase64", RegisteredBase64);
+
+                                    EmpId = cn.getUid();
+                                    Log.i("MFS_Log EmpId", EmpId);
+
+                                    Enroll_Template = Base64.decode(RegisteredBase64, Base64.DEFAULT);
+                                    Log.i("MFS_Log Enroll_Template", "" + Enroll_Template);
+
+                                    Verify_Template = new byte[fingerData.ISOTemplate().length];
+                                    System.arraycopy(fingerData.ISOTemplate(), 0, Verify_Template, 0,
+                                            fingerData.ISOTemplate().length);
+
+                                    Log.i("MFS_Log Verify_Template", "" + Verify_Template);
+
+                                    result_match = mfs100.MatchISO(Enroll_Template, Verify_Template);
+                                    Log.i("MFS_Log result_match", "" + result_match);
+
+                                    if (result_match >= 1400)
+                                    {
+                                        Log.i("MFS_Log MATCHED!!", "MATCHED!!");
+                                        makeAttendance();
+                                        break;
+                                    }
+                                    else
+                                    {
+                                        Log.i("MFS_Log NOT MATCHED!!", "NOT MATCHED!!");
+                                    }
+                                }
+                            }
+                        }
+                    }
+                    else
+                    {
+                        break;
+                    }
+                }
+
+                if (result_match < 1400)
+                {
+                    btn_signIn.setEnabled(true);
+                    btn_signOut.setEnabled(true);
+
+                    img_in_mark.setVisibility(View.GONE);
+                    img_out_mark.setVisibility(View.GONE);
+                    txt_quality_per.setText("0%");
+                    progress_quality.setProgress(0);
+                    txt_result.setText("Sorry thumb not matched");
+                    txt_result.setTextColor(getColor(R.color.RedTextColor));
+                    img_thumb_result.setImageDrawable(getDrawable(R.drawable.thumb_red));
+                    textToSpeech.speak("Sorry thumb not matched!!", TextToSpeech.QUEUE_FLUSH, null);
+                    Log.i("THUMB NOT MATCHED!!", "THUMB NOT MATCHED!!");
+
+                    final Handler handler = new Handler();
+                    handler.postDelayed(new Runnable()
+                    {
+                        @Override
+                        public void run()
+                        {
+                            txt_result.setText("");
+                            txt_quality_success.setVisibility(View.INVISIBLE);
+                            img_thumb_result.setImageDrawable(getDrawable(R.drawable.thumb_black));
+                        }
+                    }, 3000);
+                }
+            }
+        });
     }
 
     public void makeAttendance()
@@ -1301,110 +1416,6 @@ public class AttendanceActivity extends AppCompatActivity implements MFS100Event
         getDataJSON.execute();
     }
 
-    public void MatchThumbByDB(final FingerData fingerData)
-    {
-        runOnUiThread(new Runnable()
-        {
-            @Override
-            public void run()
-            {
-                List<UserDetails_Model> contacts = db.getAllContacts();
-                Log.i("MFS_Log contacts", "" + contacts);
-                result_match = 0;
-
-                /*for (UserDetails_Model cn : contacts)
-                {
-                    Log.i("thumb_reg", cn.getThumb1()+", \n"+cn.getThumb2()+
-                            ", \n"+cn.getThumb3()+", \n"+cn.getThumb4());
-                }*/
-
-                for (UserDetails_Model cn : contacts)
-                {
-                    if (result_match < 1400)
-                    {
-                        String t1 = cn.getThumb1();
-                        String t2 = cn.getThumb2();
-                        String t3 = cn.getThumb3();
-                        String t4 = cn.getThumb4();
-
-                        String thumbs[] = {t1, t2, t3, t4};
-                        Log.i("thumbs",""+  Arrays.toString(thumbs));
-
-                        for (int i = 0; i < thumbs.length; i++)
-                        {
-                            if (result_match < 1400)
-                            {
-                                RegisteredBase64 = thumbs[i];
-                                if (RegisteredBase64 != null)
-                                {
-                                    Log.i("MFS_ RegisteredBase64", RegisteredBase64);
-
-                                    EmpId = cn.getUid();
-                                    Log.i("MFS_Log EmpId", EmpId);
-
-                                    Enroll_Template = Base64.decode(RegisteredBase64, Base64.DEFAULT);
-                                    Log.i("MFS_Log Enroll_Template", "" + Enroll_Template);
-
-                                    Verify_Template = new byte[fingerData.ISOTemplate().length];
-                                    System.arraycopy(fingerData.ISOTemplate(), 0, Verify_Template, 0,
-                                            fingerData.ISOTemplate().length);
-
-                                    Log.i("MFS_Log Verify_Template", "" + Verify_Template);
-
-                                    result_match = mfs100.MatchISO(Enroll_Template, Verify_Template);
-                                    Log.i("MFS_Log result_match", "" + result_match);
-
-                                    if (result_match >= 1400)
-                                    {
-                                        Log.i("MFS_Log MATCHED!!", "MATCHED!!");
-                                        makeAttendance();
-                                        break;
-                                    }
-                                    else
-                                    {
-                                        Log.i("MFS_Log NOT MATCHED!!", "NOT MATCHED!!");
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    else
-                    {
-                        break;
-                    }
-                }
-
-                if (result_match < 1400)
-                {
-                    btn_signIn.setEnabled(true);
-                    btn_signOut.setEnabled(true);
-
-                    img_in_mark.setVisibility(View.GONE);
-                    img_out_mark.setVisibility(View.GONE);
-                    txt_quality_per.setText("0%");
-                    progress_quality.setProgress(0);
-                    txt_result.setText("Sorry thumb not matched");
-                    txt_result.setTextColor(getColor(R.color.RedTextColor));
-                    img_thumb_result.setImageDrawable(getDrawable(R.drawable.thumb_red));
-                    textToSpeech.speak("Sorry thumb not matched!!", TextToSpeech.QUEUE_FLUSH, null);
-                    Log.i("THUMB NOT MATCHED!!", "THUMB NOT MATCHED!!");
-
-                    final Handler handler = new Handler();
-                    handler.postDelayed(new Runnable()
-                    {
-                        @Override
-                        public void run()
-                        {
-                            txt_result.setText("");
-                            txt_quality_success.setVisibility(View.INVISIBLE);
-                            img_thumb_result.setImageDrawable(getDrawable(R.drawable.thumb_black));
-                        }
-                    }, 3000);
-                }
-            }
-        });
-    }
-
     protected void onStop()
     {
         UnInitScanner();
@@ -1451,5 +1462,157 @@ public class AttendanceActivity extends AppCompatActivity implements MFS100Event
         startActivity(intent);
         finish();
 
+    }
+
+    public void GetDeviceName(final String device_id)
+    {
+        class SendDeviceID extends AsyncTask<String, Void, String>
+        {
+            private URL url;
+            private String response = "";
+
+            @Override
+            protected void onPreExecute()
+            {
+                /*progressDialog1 = new ProgressDialog(AttendanceActivity.this, ProgressDialog.THEME_HOLO_LIGHT);
+                progressDialog1.setProgressStyle(ProgressDialog.STYLE_SPINNER);
+                progressDialog1.setTitle("Please wait");
+                progressDialog1.setMessage("Receiving Device ID...");
+                progressDialog1.show();*/
+            }
+
+            @Override
+            protected String doInBackground(String... params)
+            {
+                try
+                {
+                    String Transurl = ""+url_http+""+Url+"/owner/hrmapi/getnameofdeviceid/?";
+
+                    String query = String.format("android_devide_id=%s",
+                            URLEncoder.encode(device_id, "UTF-8"));
+
+                    url = new URL(Transurl + query);
+                    Log.i("url", "" + url);
+
+                    HttpURLConnection conn = (HttpURLConnection) url.openConnection();
+                    conn.setReadTimeout(10000);
+                    conn.setConnectTimeout(10000);
+                    conn.setRequestMethod("GET");
+                    conn.setDoInput(true);
+                    conn.setRequestProperty("Content-Type", "application/x-www-form-urlencoded");
+                    conn.setDoOutput(true);
+                    int responseCode = conn.getResponseCode();
+
+                    if (responseCode == HttpsURLConnection.HTTP_OK)
+                    {
+                        String line;
+                        BufferedReader br = new BufferedReader(new InputStreamReader(conn.getInputStream()));
+                        while ((line = br.readLine()) != null)
+                        {
+                            response += line;
+                        }
+                    }
+                    else {
+                        response = "";
+                    }
+                }
+                catch (SocketTimeoutException e)
+                {
+                    runOnUiThread(new Runnable() {
+                        @Override
+                        public void run() {
+                            //progressDialog1.dismiss();
+                            Toast.makeText(AttendanceActivity.this, "Slow internet / Login to captive portal", Toast.LENGTH_SHORT).show();
+                        }
+                    });
+                    Log.e("SocketTimeoutException", e.toString());
+                }
+                catch (ConnectTimeoutException e)
+                {
+                    runOnUiThread(new Runnable() {
+                        @Override
+                        public void run() {
+                            //progressDialog1.dismiss();
+                            Toast.makeText(AttendanceActivity.this, "Slow internet / Login to captive portal", Toast.LENGTH_SHORT).show();
+                        }
+                    });
+                    Log.e("ConnectTimeoutException", e.toString());
+                }
+                catch (Exception e)
+                {
+                    runOnUiThread(new Runnable() {
+                        @Override
+                        public void run() {
+                            //progressDialog1.dismiss();
+                            Toast.makeText(AttendanceActivity.this, "Slow internet / Login to captive portal", Toast.LENGTH_SHORT).show();
+                        }
+                    });
+                    Log.e("Exception", e.toString());
+                }
+
+                return response;
+            }
+
+            @Override
+            protected void onPostExecute(String result)
+            {
+                myJSON = result;
+                Log.i("response", result);
+                if (response.equals("[]"))
+                {
+                    //progressDialog1.dismiss();
+                    Toast.makeText(AttendanceActivity.this, "Sorry... Slow internet connection", Toast.LENGTH_LONG).show();
+                }
+                else
+                {
+                    try
+                    {
+                        JSONArray json = new JSONArray(result);
+                        //Log.i("json", "" + json);
+
+                        JSONObject object = json.getJSONObject(0);
+
+                        String responsecode = object.getString("responseCode");
+
+                        if (responsecode.equals("1"))
+                        {
+                            //progressDialog1.dismiss();
+
+                            TabID = object.getString("responseMessage");
+                            Log.i("TabID",TabID);
+                            txt_att_deviceid.setText("Tab"+TabID);
+                            Log.i("Device Name",txt_att_deviceid.getText().toString());
+                        }
+                        else
+                        {
+                            //progressDialog1.dismiss();
+
+                            String msg = object.getString("responseMessage");
+                            String message = msg.substring(2, msg.length()-2);
+
+                            AlertDialog.Builder alertDialog = new AlertDialog.Builder(AttendanceActivity.this, AlertDialog.THEME_DEVICE_DEFAULT_LIGHT);
+                            alertDialog.setTitle(message);
+                            alertDialog.setCancelable(true);
+                            alertDialog.setPositiveButton("Ok", new DialogInterface.OnClickListener()
+                            {
+                                @Override
+                                public void onClick(DialogInterface dialog, int which)
+                                {
+                                    dialog.dismiss();
+                                }
+                            });
+
+                            alertDialog.show();
+                        }
+                    }
+                    catch (JSONException e){
+                        //progressDialog1.dismiss();
+                        Log.i("Exception", e.toString());
+                    }
+                }
+            }
+        }
+        SendDeviceID sendid = new SendDeviceID();
+        sendid.execute();
     }
 }
